+++ src/gamecode/codelexer.h
@@ -8,7 +8,7 @@ struct Meta;
 
 struct yystype {
 	bool bValue;
-	int iValue;
+	int64_t iValue;
 	float fValue;
 	char strValue[256];
 	GenObject* metaObject;
